+++ examples/eg_patch_areas_grid.py
@@ -1,10 +1,10 @@
-# -*- coding: utf-8 -*-
 """
-Created on Fri Jan 13 12:05:46 2023
+@brief Example script for surface areas of a sphere/spheroid/ellipsoid.
+@file eg_patch_areas_grid.py
+@author Callum Marples
 
-@author: Callum Marples
-
-Examples of using the area routines.
+- Created on 13/01/2023. 
+- Last modified on 17/01/2023.
 """
 
 import leod
@@ -15,15 +15,15 @@ import os
 
 os.chdir("..")
 
-# Define shapes
+### Define shape and grid
 ell = leod.shape.EllipsoidShape(3.0, 2.0, 1.0)
-
-# Ellipsoid grid
 grid = leod.grid.Grid(181, 360)
 
+### Compute patch areas using the patch.grid routine
 p = leod.area.patch.grid(ell, grid)
-a0 = np.sum(p)
 
+# Does the combined areas of the patches agree with routines for full area?
+a0 = np.sum(p) 
 # Thomsen
 a1 = leod.area.full.thomsen(ell)
 # Legendre
@@ -31,8 +31,7 @@ a2 = leod.area.full.legendre(ell)
 # Numerical (Romberg)
 a3 = leod.area.full.numerical(ell)
 
-# North pole patch area
+# North pole patch area (should match p[0])
 an = 4.0 * leod.area.patch.ellipsoid(ell, 0.0, 0.5*grid.delta_theta, 0.0, 0.5*pi)
-
-# Pole adjacent patch area
+# Pole adjacent patch area (should be half of p[1])
 am = leod.area.patch.ellipsoid(ell, 0.5*grid.delta_theta, 1.5*grid.delta_theta, 0.0, 0.5*grid.delta_phi)